Finish making CachedArrays threadsafe (hopefully)
Minor tweak to Garbage Collection strategy.

When the `cleanup` finalizer gets called, the `Block` metadata for the
array gets appended to a work list.

During allocation (which is now a safe place from a data structure
perspective) is when we process the worklist and put all the `Blocks`
back in the heap.

Doing this DURING GC can be problematic if the GC runs during
allocation.