Fix concurrency bug.
There's potential for a CachedArray to get GC'd while eviction is
running. If that's the case, then a Block can get added to the
CacheManager's clean list that is actually in the process of being
evicted.

This adds a test for such blocks. They've already been put back in the
heap, so no work actually has to be donw.