Rework much of cached arrays.
This is an initial change set for reworking the cached arrays
implementation. Before, each array "owned" the region of data it
represented. This worked great since the pointer to the data region
could be stored inline with the CachedArray struct. However, it had some
key limitations.

We could not have multiple CachedArrays owning the same data region.
This is because maintain back edges from the cache manager to all
owners of the region for pointer swapping isn't a particularly
elegant idea. Because of this, key operations like "reshape" couldn't
return a new CachedArray of the required shape and instead needed to
return a "ReshapedArray", which isn't good for type-stability in key
places.

Furthermore, actions like locking an array required a wrapper type.

Now, managed regions of memory are implemented using a "Region" type.
Multiple arrays can share this region as it won't be garbage collected
until all objects holding it are free. Since now all users of an
allocated region go through a single point, updating the pointer will be
visible to all.

Eventually, atomic signals will have to be used to coordinate memory access and eviction.
One can imagine transitioning an array from a "Locked" state to a
"ReadOnly" state will increment a semaphore in the managed region,
keeping that region of data from being evicted until released.
Similarly, if a region of data is in the process of being moved, this
state can be indicated, keeping users from transitioning from a "Locked"
state to a modifiable state.